+++ tests/com/kaylerrenslow/armaDialogCreator/control/ControlClassInheritanceTests.java
@@ -27,6 +27,8 @@ public class ControlClassInheritanceTests {
 		tcc.extendControlClass(tcc2);
 		SerializableValue jenny = new SVInteger(867_5309);
 		if (constant.getPropertyType() != PropertyType.Int) {
+			//Need to guarantee that the value doesn't change.
+			//By differing the types, we can guarantee that there was or wasn't a change.
 			throw new IllegalStateException("propertyType for the lookup should be int for testing purposes");
 		}
 		tcc2.findRequiredProperty(constant).setValue(jenny);
@@ -68,7 +70,7 @@ public class ControlClassInheritanceTests {
 		}
 
 		//make sure this value is not equal to what the merge method returns
-		//so that we can assert that the value actually changed with the merge method!
+		//so that we can assert that the value actually changed
 		tcc2.findRequiredProperty(constant).setValue(AFont.LUCIDA_CONSOLE_B);
 
 		tcc.inheritProperty(constant);
@@ -79,7 +81,7 @@ public class ControlClassInheritanceTests {
 	}
 
 	@Test
-	public void inheritProperty_success_createTemp() throws Exception { //todo
+	public void inheritProperty_success_createTemp() throws Exception {
 		//inherit property that does not exist in the class, but does in the parent class
 		TestControlClass tcc = new TestControlClass(new TestSpecRegistry());
 		TestControlClass2 tcc2 = new TestControlClass2(new TestSpecRegistry());
@@ -97,7 +99,7 @@ public class ControlClassInheritanceTests {
 	}
 
 	@Test
-	public void inheritProperty_success_exist_wrongLookup() throws Exception { //todo
+	public void inheritProperty_success_exist_wrongLookup() throws Exception {
 		//override property that does exist in the class, but the lookups aren't equal
 
 		TestControlClass tcc = new TestControlClass(new TestSpecRegistry());
@@ -120,7 +122,8 @@ public class ControlClassInheritanceTests {
 		TestControlClass tcc = new TestControlClass(new TestSpecRegistry());
 		TestControlClass tcc2 = new TestControlClass(new TestSpecRegistry());
 		ControlPropertyLookupConstant constant = TestControlClass.requiredProperties[0];
-		SerializableValue oldValue = tcc.findProperty(constant).getValue();
+		SerializableValue oldValue = new SVString("Leroy Jenkins");
+		tcc.findProperty(constant).setValue(oldValue);
 
 		tcc.extendControlClass(tcc2);
 		SerializableValue jenny = new SVInteger(867_5309);
@@ -135,7 +138,29 @@ public class ControlClassInheritanceTests {
 	}
 
 	@Test
-	public void overrideProperty_success_doesNotExist() throws Exception { //todo
+	public void overrideProperty_success_exists_wrongLookup() throws Exception {
+		//override property that exists in the class, but lookups aren't the same
+		TestControlClass tcc = new TestControlClass(new TestSpecRegistry());
+		TestControlClass tcc2 = new TestControlClass(new TestSpecRegistry());
+		ControlPropertyLookupConstant constant = ControlPropertyLookup.IDC;
+		ControlPropertyLookupConstant constantLookalike = TestControlPropertyLookup.IDC;
+		SerializableValue oldValue = new SVString("Leroy Jenkins");
+		tcc.findProperty(constant).setValue(oldValue);
+
+		tcc.extendControlClass(tcc2);
+		SerializableValue jenny = new SVInteger(867_5309);
+		if (constant.getPropertyType() != PropertyType.Int) {
+			throw new IllegalStateException("propertyType for the lookup should be int for testing purposes");
+		}
+		tcc2.findRequiredProperty(constant).setValue(jenny);
+		tcc.inheritProperty(constant);
+		tcc.overrideProperty(constantLookalike);
+
+		assertEquals(oldValue, tcc.findRequiredProperty(constant).getValue());
+	}
+
+	@Test
+	public void overrideProperty_success_doesNotExist() throws Exception {
 		//override property that does not exist in the class, but was inherited
 		TestControlClass tcc = new TestControlClass(new TestSpecRegistry());
 		TestControlClass2 tcc2 = new TestControlClass2(new TestSpecRegistry());